refactor: update bin/npinit
— dropping flag `new`
— installed check project name in the first param position
— updated usage check (no project name given === usages() )
— flag for `dry run` can be —dry or -d
— simplified public/private repo checks
— changed flags for dependencies: —dev for devDependencies, —dep for
Dependencies. only difference is the last letter. can still be
considered as confusion but I need something short to type in and —d/—D
just didn’t sit right with me. it was a single letter with double
dashes.  ugh!!!!
— since —new was dropped, entering `test` as the project name will
generate a project name in the format `test-###` … got tired typing in
`testprojXXX` to cd into a directory.