refactor: bin/npinit -- simplify repo initialization
when github option is true and private repo is undefined or false and
project name is not `test` then go forth with git init, create remote
and push. if either repo option or project name is `test` along with
the github option, then a private repo only will be created.